Fix clippy, lint
